+++ CHANGELOG.md
@@ -10,8 +10,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Added
 
-- added `delete` method to `FREDJob` class.
-- added a method to FREDRun, get_table_variable, that loads table varibale out for a FRED run.
+- added `delete_snapshots()` method to `FREDJob` class.
+- added `delet()` method to `Snapshot` class.
+- added `delete()` method to `FREDJob` class.
+- added a method to `FREDRun`, `get_table_variable`.
 
 ### Fixed